split large modules into focused helpers
