+++ src/gengowatcher/ui_textual.py
@@ -34,12 +34,14 @@ from textual.widgets import (
 from .config import AppConfig
 from .state import AppState
 from .stats import StatsManager
+from .ui_charts import (
+    BAR_CHARS,
+    _aggregate_series,
+    _render_chart,
+    _render_chart_with_axes,
+    _render_plotext_bar_chart,
+)
 from .watcher import TIER_UNIT_RATES, GengoWatcher
-
-try:
-    import plotext as plotext
-except ImportError:  # pragma: no cover - optional runtime dependency
-    plotext = None
 
 # =============================================================================
 # Constants
@@ -373,158 +375,6 @@ def _derive_display_word_count(job: dict[str, Any]) -> int:
         rate = TIER_UNIT_RATES["standard"]
 
     return max(1, int(round(reward / rate)))
-
-
-# Fractional block characters for bar chart rendering
-# Characters arranged from empty to full: ▁▂▃▄▅▆▇█
-BAR_CHARS = " ▁▂▃▄▅▆▇█"
-
-
-def _render_chart(
-    values: list[float],
-    width: int = 20,
-    height: int = 5,
-) -> str:
-    """
-    Render a bar chart using fractional block characters.
-
-    Args:
-        values: List of numeric values to display
-        width: Width of the chart in characters
-        height: Height of the chart in lines
-
-    Returns:
-        String representation of the chart with newlines
-    """
-    if not values or width <= 0 or height <= 0:
-        return ""
-
-    # Normalize values to fit within the height
-    max_val = max(values) if values else 1.0
-    if max_val == 0:
-        max_val = 1.0
-
-    # Resample values to fit width if needed
-    if len(values) > width:
-        # Downsample by averaging buckets
-        step = len(values) / width
-        resampled = []
-        for i in range(width):
-            start_idx = int(i * step)
-            end_idx = int((i + 1) * step)
-            bucket = values[start_idx:end_idx]
-            resampled.append(sum(bucket) / len(bucket) if bucket else 0)
-        values = resampled
-    elif len(values) < width:
-        # Pad with zeros on the right
-        values = list(values) + [0.0] * (width - len(values))
-
-    # Normalize to chart height (using fractional blocks)
-    # Each position can be 0 to (height * 8) where 8 is the number of
-    # fractional states
-    max_units = height * 8
-    normalized = [(v / max_val) * max_units for v in values]
-
-    # Build chart from top to bottom
-    lines = []
-    for row in range(height - 1, -1, -1):
-        line = ""
-        for col_val in normalized:
-            # Determine which character to use for this row
-            # row represents height from bottom (0 = bottom row, height-1 = top
-            # row)
-            units_at_col = col_val
-            units_needed_for_row = row * 8
-
-            if units_at_col > units_needed_for_row + 8:
-                # Full block for this row
-                line += BAR_CHARS[-1]  # █
-            elif units_at_col > units_needed_for_row:
-                # Partial block for this row
-                fraction = int(units_at_col - units_needed_for_row)
-                line += BAR_CHARS[min(fraction, len(BAR_CHARS) - 1)]
-            else:
-                # Empty for this row
-                line += BAR_CHARS[0]  # space
-        lines.append(line)
-
-    return "\n".join(lines)
-
-
-def _aggregate_series(values: list[float], bin_size: int = 2) -> list[float]:
-    """Aggregate a series into fixed-size bins."""
-    if bin_size <= 1:
-        return list(values)
-    aggregated: list[float] = []
-    for i in range(0, len(values), bin_size):
-        aggregated.append(float(sum(values[i : i + bin_size])))
-    return aggregated
-
-
-def _render_chart_with_axes(
-    values: list[float],
-    *,
-    width: int = 12,
-    height: int = 4,
-    x_left: str = "old",
-    x_right: str = "new",
-) -> str:
-    """Render chart with minimal y-axis and x-axis labels."""
-    chart = _render_chart(values, width=width, height=height)
-    if not chart:
-        return ""
-
-    lines = chart.splitlines()
-    max_val = max(values) if values else 0.0
-    if max_val <= 0:
-        max_val = 1.0
-
-    y_label_width = max(1, len(str(int(round(max_val)))))
-    with_axis: list[str] = []
-    for row_idx, line in enumerate(lines):
-        # Top row shows the highest approximate bucket value.
-        approx_value = int(round(max_val * (height - row_idx) / height))
-        with_axis.append(f"{approx_value:>{y_label_width}} |{line}")
-
-    with_axis.append(f"{0:>{y_label_width}} +{'─' * width}")
-    left_pad = " " * (y_label_width + 2)
-    spacing = max(1, width - len(x_left) - len(x_right))
-    with_axis.append(f"{left_pad}{x_left}{' ' * spacing}{x_right}")
-    return "\n".join(with_axis)
-
-
-def _render_plotext_bar_chart(
-    values: list[float],
-    *,
-    width: int,
-    height: int,
-    x_left: str,
-    x_mid: str,
-    x_right: str,
-) -> str:
-    """Render a bar chart via plotext with true axes/ticks."""
-    if plotext is None or not values or width <= 0 or height <= 0:
-        return ""
-
-    try:
-        x = list(range(1, len(values) + 1))
-        mid = max(1, len(values) // 2)
-        right = len(values)
-
-        plotext.clear_figure()
-        plotext.plotsize(width, height)
-        plotext.bar(x, values, fill=True, width=0.8)
-        plotext.xticks([1, mid, right], [x_left, x_mid, x_right])
-        plotext.ylabel("jobs")
-        plotext.xlabel("time")
-        plotext.grid(True)
-
-        # Convert ANSI output into plain text, preserving chart glyphs.
-        built = plotext.build()
-        plotext.clear_figure()
-        return str(Text.from_ansi(built)).rstrip()
-    except Exception:
-        return ""
 
 
 def _parse_job_title_fallback(title: Any) -> tuple[str, str]:
